% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaXTalk-package.R
\docType{package}
\name{SpaXTalk-package}
\alias{SpaXTalk-package}
\title{SpaXTalk: R package for identifying and quantifying crosstalk pathways using spatial trancriptomics data}
\description{
The R package to analyze potential crosstalk between (1) multiple regulatory pathways or (2) multiple sender cells induced by cell-cell communication (CCC) using spatial transcriptomics data.

SpaXTalkR provides tools for quantifying downstream effects of
cell-cell communication signals and how they vary across space
}
\author{
\strong{Maintainer}: Jiawen Hou \email{lithiumhou@gmail.com}

Authors:
\itemize{
  \item Jiawen Hou \email{lithiumhou@gmail.com}
}

}
\keyword{internal}
